add(sysinfo): add

+++ kernel/sysinfo.h
@@ -1,5 +1,5 @@
 struct sysinfo{
-    int freemem; // number of bytes of free memory
-    int nproc; // number of processes whose state is not UNUSED
-    int nopenfiles; // number of opening files in whole system
+    uint64 freemem; // number of bytes of free memory
+    uint64 nproc; // number of processes whose state is not UNUSED
+    uint64 nopenfiles; // number of opening files in whole system
 };
